Juce 6; Submodules; Deprecation; and OSes
This diff does a couple of things

1. Remove the copied VST2SDK and Maximillian libraries
2. Add JUCE6 and Maximillian as git submodules
3. Create a CMakeFile which builds the assets with JUCE6 CMake
4. Removes the various derprecation warnings for JUCE6
5. Confirms the resulting synth works on mac linux and win
   as standalone VST3 and (mac) AU
6. Modifies the .jucer file to still allow Projucer based
   workflow with the new submodules

I note the codebase doesn't have a license. To use JUCE in this
way I think you may need to be GPL3 or have a JUCE license.

And feel free to merge this or not. I was just helping a collaborator
who wanted your synth on linux! Hope it helps!

s