Fix a pop and click in chorus
Chorus did a delay bin sin(phase) but the delay did a fabs which
meant correctly that it was always looking in the past, but that
the delay value changed very rapidly near zero, leading to occasional
pops and clicks. Repalce 350 * sin with 175 * ( 1 + sin ) + 1 in the delay
time to stop this.